Upfate tabSize usage in query result. Fixes #1504
