+++ packages/altair-app/src/app/modules/altair/altair.module.ts
@@ -83,7 +83,6 @@ const providers = [
   services.ThemeRegistryService,
   services.SubscriptionProviderRegistryService,
   services.PluginContextService,
-  services.BackupService,
   // Setting the reducer provider in main.ts now (for proper config initialization)
   // reducerProvider,
   CookieService,
