+++ packages/altair-app/src/app/modules/altair/services/index.ts
@@ -15,4 +15,3 @@ export { PreRequestService } from './pre-request/pre-request.service';
 export { ThemeRegistryService } from './theme/theme-registry.service';
 export { SubscriptionProviderRegistryService } from './subscriptions/subscription-provider-registry.service';
 export { PluginContextService } from './plugin/context/plugin-context.service';
-export { BackupService } from './backup/backup.service';
